Extract tip button creating from twitter content script
